+++ tests/test-helper.js
@@ -1,28 +1,21 @@
+/* globals require, mocha */
 import resolver from './helpers/resolver';
-import { setResolver } from 'ember-qunit';
+import { setResolver } from 'ember-mocha';
 
 setResolver(resolver);
 
 document.write('<div id="ember-testing-container"><div id="ember-testing"></div></div>');
 
-Ember.Test.adapter = Ember.Test.MochaAdapter.create();
 $(document).ready(function(){
-  mocha.checkLeaks();
-  mocha.globals([
-    'visit',
-    'click',
-    'keyEvent',
-    'fillIn',
-    'find',
-    'findWithAssert',
-    'wait',
-    'andThen',
-    'triggerEvent',
-    'LiveReload',
-    'currentRouteName',
-    'currentPath',
-    'currentURL',
-    '__PROMISE_INSTRUMENTATION__'
-  ]);
+  // Rename elements from qunit -> mocha
+  $('#qunit').attr('id', 'mocha');
+  $('#qunit-fixture').attr('id', 'mocha-fixture');
+
+  // Declare `expect` as a global here instead of as a var in individual tests.
+  // This avoids jshint warnings re: `Redefinition of 'expect'`.
+  window.expect = chai.expect;
+
+  require('ember-cli/test-loader')['default'].load();
+
   mocha.run();
 });